% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_lmer_coefficients.R
\name{aggregate_lmer_coefficients}
\alias{aggregate_lmer_coefficients}
\title{Function to Aggregate Results of `lmer_modeling` function.}
\usage{
aggregate_lmer_coefficients(
  save_model_directory,
  model_names_pattern,
  save_model_path
)
}
\arguments{
\item{save_model_directory}{directory of saved models.}

\item{model_names_pattern}{model name or pattern of models name.}

\item{save_model_path}{aggregated models save path.}
}
\value{

}
\description{
This function aggregate coefficients of lmer models per each SNP
}
