% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_lmer_modeling.r, R/lmer_modeling.r
\name{lmer_modeling}
\alias{lmer_modeling}
\title{LMER Modeling Function}
\usage{
lmer_modeling(data_path, simulation_name, formula, save_model_path)

lmer_modeling(data_path, simulation_name, formula, save_model_path)
}
\arguments{
\item{data_path}{the path of partition data}

\item{simulation_name}{when save output models, is used in name of save model file.}

\item{formula}{Character variable for `lmer` formula.}

\item{save_model_path}{The directory of saving model outout.}
}
\value{



}
\description{
This function is designed to run `lmer` models on data partition.
It takes a formula based on variables that are set in function
`phonotype_gwas_partitioner` (clinical and PCA variables).
To set WGS in the `lmer` formula, use `copy_number` term.

This function is designed to run `lmer` models on data partition.
It takes a formula based on variables that are set in function
`wgs_adni_partitioner` (clinical and PCA variables).
To set WGS in the `lmer` formula, use `copy_number` term.
}
