% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_rscripts.r
\name{parallel_rscripts}
\alias{parallel_rscripts}
\title{Run R Scripts Parallel with System Control}
\usage{
parallel_rscripts(
  rscript_path = "/data/array/mshoai/ADprogressionData/ADNI/scripts/data_modeling.r",
  args,
  core_numbers = 50,
  free_memory_treshold = 75,
  free_cpu_treshold = 75,
  sleep_time = 10
)
}
\arguments{
\item{rscript_path}{path of rscript.}

\item{args}{input arguments of rscript.}

\item{core_numbers}{number of cores to run parallel jobs}

\item{free_memory_treshold}{upper bound for memory usage}

\item{free_cpu_treshold}{upper bound for CPU usage}

\item{sleep_time}{sleep time between two jobs in seconds}
}
\value{

}
\description{
Run R Scripts Parallel with System Control
}
