% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonotype_gwas_partitioner.R
\name{phonotype_gwas_partitioner}
\alias{phonotype_gwas_partitioner}
\title{Create Data Partitions for LMER Modeling}
\usage{
phonotype_gwas_partitioner(
  pheno_data,
  geno_data,
  joinBy,
  partition_number,
  partitions_save_path
)
}
\arguments{
\item{pheno_data}{phenotype data.frame}

\item{geno_data}{GWAS data frame that contains one ID and SNP columns}

\item{joinBy}{The character value joinBy is common between pheno_data and geno_data columns.}

\item{partition_number}{is an integer number that represents the number of data is divided.}

\item{partitions_save_path}{is the save path of data partitions.}
}
\value{

}
\description{
This function creates data partitions for `lmer` modeling.
To create data, first we load adnimerge and select specific columns.
Then we load the WGS PCA data and select the required columns.
In the next step, merge the two datasets into one.
Finally, we merge WGS data by `PTID`. Data is melted and created into a
tidy version before being used in data.table format.
}
