% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCF2Delata.R
\name{VCF2Delta}
\alias{VCF2Delta}
\title{Read VCF Data and Convert to Delta Format}
\usage{
VCF2Delta(vcfPath, savePath, cores = 4)
}
\arguments{
\item{vcfPath}{path of VCF data}

\item{savePath}{save path for Delta file}

\item{cores}{Spark local number of cores}
}
\value{

}
\description{
Read VCF Data and Convert to Delta Format
}
