% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_to_Rscripts.R
\name{function_to_Rscript}
\alias{function_to_Rscript}
\title{Convert R Functions to Scripts that are given arguments from the Console}
\usage{
function_to_Rscript(
  script_name = "Parallel_Modeling.R",
  function_from_source = NULL,
  function_name = "lmer_modeling",
  packages = c("IRISUtils"),
  arguments = c("data_path", "simulation_name", "formula", "save_model_path"),
  arguments_class = c("character", "character", "character", "character")
)
}
\arguments{
\item{script_name}{path of generated script.}

\item{function_from_source}{the path of function that add function from source.}

\item{function_name}{name of function to run in script.}

\item{packages}{A vector of package names that have been loaded into the script.}

\item{arguments}{A vector of arguments are given from console.}

\item{arguments_class}{A vector of input arguments types.}
}
\value{

}
\description{
Convert R Functions to Scripts that are given arguments from the Console
}
