% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_rscripts.r
\name{parallel_rscripts}
\alias{parallel_rscripts}
\title{Run R Scripts Parallel with System Control}
\usage{
parallel_rscripts(
  rscript_path,
  args,
  free_memory_treshold = 75,
  free_cpu_treshold = 75,
  sleep_time = 10,
  swap = 8,
  process = F
)
}
\arguments{
\item{rscript_path}{path of rscript.}

\item{args}{input arguments of rscript.}

\item{free_memory_treshold}{upper bound for memory usage}

\item{free_cpu_treshold}{upper bound for CPU usage}

\item{sleep_time}{sleep time between two jobs in seconds}

\item{swap}{server swap memory in GB}

\item{process}{Boolean variable to see function messages}
}
\value{

}
\description{
Run R Scripts Parallel with System Control
}
